modify

+++ services/audiomanager/servicesource/src/daudio_ipc_callback_proxy.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2023 Huawei Device Co., Ltd.
+ * Copyright (c) 2022 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -16,7 +16,6 @@
 #include "daudio_ipc_callback_proxy.h"
 
 #include "daudio_errorcode.h"
-#include "daudio_ipc_interface_code.h"
 
 #undef DH_LOG_TAG
 #define DH_LOG_TAG "DAudioIpcCallbackProxy"
@@ -59,7 +58,7 @@ int32_t DAudioIpcCallbackProxy::OnNotifyUnregResult(const std::string &devId, co
         return ERR_DH_AUDIO_SA_WRITE_PARAM_FAIED;
     }
 
-    Remote()->SendRequest(static_cast<uint32_t>(NOTIFY_UNREGRESULT), data, reply, option);
+    Remote()->SendRequest(NOTIFY_UNREGRESULT, data, reply, option);
     int32_t ret = reply.ReadInt32();
     return ret;
 }
